Добавляет редактирование точки.
