+++ src/render.js
@@ -13,7 +13,7 @@ const createElement = (template) => {
 };
 
 const render = (component, container, place = RenderPosition.BEFOREEND) => {
-  container.insertAdjacentElement(place, component.getElement());
+  container.insertAdjacentElement(place, component.element);
 };
 
 export {RenderPosition, createElement, render};
