+++ src/view/filter-view.js
@@ -1,4 +1,6 @@
-import {createElement} from '../render.js';
+import {
+  createElement
+} from '../render.js';
 
 const createFilterTemplate = () => (
   ` <form class="trip-filters" action="#" method="get">
@@ -6,28 +8,32 @@ const createFilterTemplate = () => (
         <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
         <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
       </div>
+
       <div class="trip-filters__filter">
         <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
         <label class="trip-filters__filter-label" for="filter-future">Future</label>
       </div>
+
       <button class="visually-hidden" type="submit">Accept filter</button>
     </form>`
 );
 
 export default class FilterView {
-  getTemplate() {
+  #element = null;
+
+  get template() {
     return createFilterTemplate();
   }
 
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
 
-    return this.element;
+    return this.#element;
   }
 
   removeElement() {
-    this.element = null;
+    this.#element = null;
   }
 }
