+++ src/view/sort-view.js
@@ -1,4 +1,6 @@
-import {createElement} from '../render.js';
+import {
+  createElement
+} from '../render.js';
 
 const createSortTemplate = () => (
   ` <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
@@ -6,18 +8,22 @@ const createSortTemplate = () => (
         <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day">
         <label class="trip-sort__btn" for="sort-day">Day</label>
       </div>
+
       <div class="trip-sort__item  trip-sort__item--event">
         <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
         <label class="trip-sort__btn" for="sort-event">Event</label>
       </div>
+
       <div class="trip-sort__item  trip-sort__item--time">
         <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time" disabled>
         <label class="trip-sort__btn" for="sort-time">Time</label>
       </div>
+
       <div class="trip-sort__item  trip-sort__item--price">
         <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price" checked>
         <label class="trip-sort__btn" for="sort-price">Price</label>
       </div>
+
       <div class="trip-sort__item  trip-sort__item--offer">
         <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
         <label class="trip-sort__btn" for="sort-offer">Offers</label>
@@ -26,19 +32,21 @@ const createSortTemplate = () => (
 );
 
 export default class SortView {
-  getTemplate() {
+  #element = null;
+
+  get template() {
     return createSortTemplate();
   }
 
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
 
-    return this.element;
+    return this.#element;
   }
 
   removeElement() {
-    this.element = null;
+    this.#element = null;
   }
 }
